+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from "./chain.js";
 export * from "./bridge.js";
 export * from "./dex.js";
+export * from "./feeds.js";
 export * from "./logs.js";
 export * from "./markethours.js";
 export * from "./retryfetch.js";